third_party/blink/web_tests/fast/filesystem/op-remove-expected.txt

* Running: RemoveSimple
PASS Succeeded: "/a".remove()
PASS Succeeded: "/b".remove()
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry does NOT exist: /a
PASS Verified entry does NOT exist: /b
* Running: RemoveNonRecursiveWithChildren
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/b name:b type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/c name:c type:FILE}
* Running: RemoveRecursiveWithoutChildren
PASS Succeeded: "/a".removeRecursively()
PASS Verified entry does NOT exist: /a
* Running: RemoveRecursiveWithChildren
PASS Succeeded: "/a".removeRecursively()
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry does NOT exist: /a
Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
